pyproject.toml
[build-system]
requires = ["setuptools>=55", "wheel"]
build-backend = "setuptools.build_meta"

# [build-system]
# requires = ["poetry-core"]
# build-backend = "poetry.core.masonry.api"

[project]
name = "hwpccalc"
version = "1.0.5b1"
dependencies=[
    "aiobotocore",
    "bokeh==2.4.3",
    "boto3[crt]",
    "dask[distributed]",
    "dask-cloudprovider[aws]",
    "distributed",
    "lz4",
    "python-dotenv",
    "scipy",
    "tornado==6.1",
    "xarray",
]

# [tool.setuptools]
# packages = {"hwpccalc"}
# package-dir = {hwpccalc = "src/hwpccalc"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.black]
line-length = 150
target-version = ['py310']
include = '\.pyi?$'
# exclude = """ 
# {
#     __pycache__,
#     ./.venv*
# }
# """
# 'extend-exclude' excludes files or directories in addition to the defaults
# extend-exclude = "hwpc\\/model_data_xr\\.py"

# [tool.poetry]
# name = "hwpccalc"
# version = "1.0.1"
# description = "The Harvested Wood Products Carbon Calculator."
# authors = ["Robb Lankston <robb@lankstonconsulting.com>"]
# readme = "README.md"
# packages = [{include = "hwpccalc"}]

# [tool.poetry.dependencies]
# python = "^3.10"
# bokeh = "^2.4"
# boto3 = {extras = ["crt"], version = "^1.26"}
# dask = {extras = ["distributed"], version = "2022.11"}
# dask-cloudprovider = {extras = ["aws"], version = "*"}
# distributed = "*"
# lz4 = "^4"
# python-dotenv = "*"
# scipy = "*"
# xarray = ">=2022.11"

# [tool.poetry.group.dev]
# optional = true

# [tool.poetry.group.dev.dependencies]
# build = "*"
# wheel = "*"
# awscli = "*"

# [tool.poetry.group.test]
# optional = true

# [tool.poetry.group.test.dependencies]
# plotly = "*"